% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_internals.R
\name{FindandCollect_airpres}
\alias{FindandCollect_airpres}
\title{Internal functions}
\usage{
FindandCollect_airpres(lat, long, start_datetime, end_datetime)
}
\description{
Not intended to be called directly by the user.
}
\details{
Not intended to be called directly by the user.
}
\keyword{internal}
