% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASE_funcs.R
\name{fit_BASE}
\alias{fit_BASE}
\title{Internal functions}
\usage{
fit_BASE(directory, interval = 900, n.iter = 10000, n.burnin = 5000)
}
\description{
Not intended to be called directly by the user.
}
\details{
Not intended to be called directly by the user.
}
\keyword{internal}
