% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_metabolism.R
\name{fit_metabolism}
\alias{fit_metabolism}
\title{Fit stream metabolism models}
\usage{
fit_metabolism(fitdata)
}
\arguments{
\item{fitdata}{the output of \link{prep_metabolism}.}
}
\value{
returns the output of \code{streamMetabolizer}'s \code{metab}
  function.
}
\description{
Passes data and model parameters to \code{streamMetabolizer} or
\code{BASE} according to arguments passed to \link{prep_metabolism}.
NOTE: support for modeling with
\code{BASE} is currently in development. Please use \code{streamMetabolizer}
in the meantime.
}
\details{
This function supports only default streamMetabolizer parameters at the
moment. Namely, \code{err_obs_iid=TRUE, err_proc_acor=FALSE,}
\code{ode_method='trapezoid', deficit_src='DO_mod'}.
If \code{prep_metabolism}'s
\code{model_type} argument is set to \code{'mle'}, then the following
defaults will also be set: \code{engine='nlm', pool_K600='none',}
\code{proc_err=FALSE}. If \code{model_type} is set to \code{'bayes'},
these will
instead be set to \code{engine='stan', pool_K600='binned', proc_err=TRUE},
and \code{K600_lnQ_nodes_centers} will be a sequence of length 7 from
the minimum to the maximum natural log of daily discharge.

Until this function is made more flexible, you can adjust the above
parameters by using \code{streamMetabolizer}'s functions directly.
See the example section below for more.
}
\examples{
streampulse_data = request_data(sitecode='NC_Eno',
    startdate='2016-06-10', enddate='2016-10-23')

fitdata = prep_metabolism(d=streampulse_data, type='bayes',
    model='streamMetabolizer', interval='15 min',
    rm_flagged=list('Bad Data', 'Questionable'), fillgaps=fillgaps,
    zq_curve=list(sensor_height=NULL, Z=Z_data, Q=Q_data,
    fit='power', plot=TRUE), estimate_areal_depth=TRUE)

## fit model with default parameters
modelfit = fit_metabolism(fitdata)

## fit model with custom parameters
class(fitdata) = 'data.frame' #just a formality, execute and disregard
modname = mm_name(type='bayes', pool_K600='binned',
    err_obs_iid=TRUE, err_proc_acor=FALSE, err_proc_iid=TRUE,
    ode_method = 'trapezoid', deficit_src='DO_mod', engine='stan')
modspecs = specs(modname)

#overwrite default ln(Q) node centers
addis = tapply(log(fitdata$discharge), substr(fitdata$solar.time,1,10), mean)
modspecs$K600_lnQ_nodes_centers = seq(from=min(addis),
    to=max(addis), length.out=7)

modelfit = metab(specs=modspecs, data=fitdata)
}
\seealso{
\code{\link{request_data}} for acquiring StreamPULSE data;
  \code{\link{prep_metabolism}} for organizing data and acquiring additional
  variables.
}
\author{
Mike Vlah, \email{vlahm13@gmail.com}

Aaron Berdanier
}
