% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_metabolism.R
\name{predict_metabolism}
\alias{predict_metabolism}
\title{Predict metabolism from a fitted model}
\usage{
predict_metabolism(model_fit)
}
\arguments{
\item{model_fit}{the output of \link{fit_metabolism}.}
}
\value{
returns the output of \code{streamMetabolizer}'s \code{predict_metab}
  function.
}
\description{
A thin wrapper for \code{streamMetabolizer}'s \code{predict_metab} function.
Generates estimates of GPP, ER, and K600.
}
\details{
See documentation for \code{streamMetabolizer::predict_metab}.
}
\examples{
streampulse_data = request_data(sitecode='NC_Eno',
    startdate='2016-06-10', enddate='2016-10-23')

fitdata = prep_metabolism(d=streampulse_data, type='bayes',
    model='streamMetabolizer', interval='15 min',
    rm_flagged=list('Bad Data', 'Questionable'), fillgaps=fillgaps,
    zq_curve=list(sensor_height=NULL, Z=Z_data, Q=Q_data,
    fit='power', plot=TRUE), estimate_areal_depth=TRUE)

modelfit = fit_metabolism(fitdata)

predictions = predict_metabolism(modelfit)
}
\seealso{
\code{\link{request_data}} for acquiring StreamPULSE data;
  \code{\link{prep_metabolism}} for organizing data and acquiring additional
  variables; \code{\link{fit_metabolism}} for fitting models.
}
\author{
Mike Vlah, \email{vlahm13@gmail.com}

Aaron Berdanier
}
