% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_available_results.R
\name{query_available_results}
\alias{query_available_results}
\title{View metabolism model results available on the StreamPULSE server.}
\usage{
query_available_results(region = NULL, site = NULL, year = NULL)
}
\arguments{
\item{region}{string representing region. Must be a StreamPULSE region
abbreviation like 'NC' or 'AZ', or 'all' for a list of all available model results.}

\item{site}{string representing site. Must be a StreamPULSE site
abbreviation like 'Eno' or 'LV'.}

\item{year}{string or numeric representing year, e.g. '2015' or 2015.}
}
\value{
returns a list containing a single data frame with region, site,
  and year columns.
}
\description{
Uses StreamPULSE API to query data. To download model results, see
\code{request_results}.
}
\details{
The data returned by this function depend on which of the input parameters
are specified and which are omitted. See examples for all possible ways
to use this function.
}
\examples{
View all available model results from region
query_available_results(region='NC')

View all available model results from all regions
query_available_results(region='all')

View all available model results from site
query_available_results(region='NC', site='Eno')

View all available model results from year
query_available_results(year=2017)

View all available model results from region and year
query_available_results(region='NC', year=2017)
}
\seealso{
\code{\link{request_results}} for downloading StreamPULSE model
  outputs.
}
\author{
Mike Vlah, \email{vlahm13@gmail.com}
}
