% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_data.R
\name{request_data}
\alias{request_data}
\title{Retrieve data from the StreamPULSE database}
\usage{
request_data(sitecode, startdate = NULL, enddate = NULL, variables = NULL,
  flags = TRUE, token = NULL)
}
\arguments{
\item{sitecode}{underscore-separated region and site code, e.g. 'NC_Eno'.
Full list of regions and site codes available at
\url{https://data.streampulse.org/sitelist}.}

\item{startdate}{date string formatted 'YYYY-MM-DD', representing the first
day of data to be
requested. All available records for the specified day will be requested.
If data coverage does not extend this far in time, the argument
will be adjusted to the first day with available data. Omit this argument
to include all records back to the first available. To see full available
date range, select a site from the dropdown at
\url{https://data.streampulse.org/visualize}. This will populate the date
input field with the first and last available dates.}

\item{enddate}{date string formatted 'YYYY-MM-DD', representing the last day of
data to be
requested. All available records for the specified day will be requested.
If data coverage does not extend this far in time, the argument
will be adjusted to the last day with available data. Omit this argument to
include all records up to the last available. To see full available
date range, select a site from the dropdown at
\url{https://data.streampulse.org/visualize}. This will populate the date
input field with the first and last available dates.}

\item{variables}{character vector of variable names to request. To see which
variables are available for a given site, select a site from the dropdown
at \url{https://data.streampulse.org/visualize}. Omit this argument to
request all variables potentially useful for metabolism modeling.}

\item{flags}{logical. If \code{TRUE}, information about data points flagged
as interesting, questionable, or "bad data" will be included. If
\code{FALSE}, data will still be included, but without accompanying flag
information (There's no good reason to omit flag information, and this
parameter will eventually be deprecated.).}

\item{token}{a unique alphanumeric string for each registered user of
StreamPULSE. Only necessary for accessing embargoed data.}
}
\value{
returns a \code{list} containing two \code{data.frame}s.
  The first contains all requested
  data. The second contains site metadata. Variable names and corresponding
  values are stacked in two long columns. Use \code{prep_metabolism} to
  format the output of this function.
}
\description{
Uses StreamPULSE API to query data via MySQL. Visit the data portal at
\url{https://data.streampulse.org} to download data via a browser.
}
\details{
Request data for a single region and site within the StreamPULSE database.
All sites are embargoed for 1 year or more from the time they first appear
in the database. Embargoed data are only available to the user who submits
them, and can be accessed via a unique token.
}
\examples{
streampulse_data = request_data(sitecode='NC_Eno',
    startdate='2016-06-10', enddate='2016-10-23')
}
\seealso{
\code{\link{prep_metabolism}} for organizing data returned by this
  function.
}
\author{
Mike Vlah, \email{vlahm13@gmail.com}

Aaron Berdanier
}
