% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapfill_funcs.R
\name{series_impute}
\alias{series_impute}
\title{Internal functions}
\usage{
series_impute(x, tol, samp, algorithm, variable_name, ...)
}
\description{
Not intended to be called directly by the user.
}
\details{
Not intended to be called directly by the user.
}
\keyword{internal}
