% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_results.R
\name{request_results}
\alias{request_results}
\title{Retrieve model results from the StreamPULSE server}
\usage{
request_results(sitecode, year, token = NULL)
}
\arguments{
\item{sitecode}{underscore-separated region and site code, e.g. 'NC_Eno'.
Full list of regions and site codes available at
\url{https://data.streampulse.org/sitelist}. Or, you can use the
\code{query_available_data} function in this package.}

\item{year}{string or numeric representing year, e.g. '2015' or 2015.}

\item{token}{a unique alphanumeric string for each registered user of
StreamPULSE. Only necessary for accessing embargoed results. Email
StreamPULSE developer Mike Vlah (\email{vlahm13@gmail.com}) to receive
your token.}
}
\value{
returns a list containing an unfortunately limited collection of
  information about the "best" model result we have on record for the
  site and year requested. If the output of \code{streamMetabolizer}'s
  \code{metab}
  function is a variable called \code{x}, this function will give you x@fit,
  x@data, and x@data_daily. Likewise, if the output of \code{StreamPULSE}'s
  fit_metabolism is a variable called \code{y}, this function will give you
  y$fit@fit, y$fit@data, and y$fit@data_daily. In the future, this function
  will return a lot more relevant information, but at the moment only these
  three list elements are stored on our server for each "best" model run.

  Model "bestness" is determined by an automatic
  comparison of five criteria each time a new model is fit: 1) proportion of
  daily GPP estimates that go negative; 2) proportion of daily ER estimates
  that go positive; 3) correlation between daily ER and K600;
  4) maximum daily K600; and 5) temporal coverage. A score is assigned
  for each of these criteria, and an aggregate score is determined.
}
\description{
Uses StreamPULSE API to retrieve \code{.rds} objects containing the outputs
of streamMetabolizer
runs for specific siteyears. To view available model results, see
\code{query_available_results}. Visit the data portal at
\url{http://data.streampulse.org:3838/streampulse_diagnostic_plots/}
to visualize model results via a browser.
}
\details{
Request results for a single region, site, and year.
Some sites are "embargoed," meaning results from those sites are
kept private and can only be accessed by authorized users.
If you are authorized, you can access your embargoed results using
a unique token, which currently you can only receive by emailing StreamPULSE
developer Mike Vlah (\email{vlahm13@gmail.com}).
}
\examples{
res = query_available_results(sitecode='FL_NR1000', year=2016)
}
\seealso{
\code{\link{query_available_results}} for determining which models
 are available for download.
}
\author{
Mike Vlah, \email{vlahm13@gmail.com}
}
